+++ lib/src/wrap.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/widgets.dart';
 
 /// An encapsulation of [Builder], also used as an alias.
@@ -49,4 +50,65 @@ extension WrapAlign on Widget {
   /// Wrap current widget with [Center],
   /// which will display current widget in the center of the parent.
   Widget get center => Center(child: this);
+}
+
+extension WrapPadding on Widget {
+  /// Wrap current widget with [Padding].
+  /// And it's parameters are encapsulations for [EdgeInsets.only].
+  Widget padding({
+    double top = 0,
+    double right = 0,
+    double bottom = 0,
+    double left = 0,
+  }) =>
+      Padding(
+        padding: EdgeInsets.only(
+          top: top,
+          right: right,
+          bottom: bottom,
+          left: left,
+        ),
+        child: this,
+      );
+
+  /// Wrap current widget with [Padding].
+  /// And it's parameters are encapsulations for [EdgeInsets.all].
+  Widget paddingAll(double value) =>
+      Padding(padding: EdgeInsets.all(value), child: this);
+
+  /// Wrap current widget with [Padding].
+  /// And it's parameters are encapsulations for [EdgeInsets.symmetric].
+  Widget paddingSymmetric({
+    double vertical = 0,
+    double horizontal = 0,
+  }) =>
+      Padding(
+        padding: EdgeInsets.symmetric(
+          vertical: vertical,
+          horizontal: horizontal,
+        ),
+        child: this,
+      );
+}
+
+extension WrapList on List<Widget> {
+  Widget get asColumn => Column(mainAxisSize: MainAxisSize.min, children: this);
+
+  Widget get asRow => Row(mainAxisSize: MainAxisSize.min, children: this);
+
+  Widget get asStack => Stack(children: this);
+}
+
+extension WrapGesture on Widget {
+  Widget on({
+    void Function()? tap,
+    void Function()? doubleTap,
+    void Function()? longPress,
+  }) =>
+      GestureDetector(
+        onTap: tap,
+        onDoubleTap: doubleTap,
+        onLongPress: longPress,
+        child: this,
+      );
 }